// client/shared/src/schema.ts
export type ID = string
export type GitObjectID = string
export type DateTime = string
export type JSONCString = string
// ...
export interface IBatchChangeConnection {
    __typename: 'BatchChangeConnection'

    /**
     * A list of batch changes.
     */
    nodes: IBatchChange[]

    /**
     * The total number of batch changes in the connection.
     */
    totalCount: number

    /**
     * Pagination information.
     */
    pageInfo: IPageInfo
}
// ...
export interface IExecutorSecret {
    __typename: 'ExecutorSecret'

    /**
     * The unique identifier of the secret.
     */
    id: ID

    /**
     * The key under which the secret is available. Secrets are usually exposed
     * as environment variables named using this key.
     * Recommended format: uppercase letters, numbers and underscores.
     */
    key: string

    /**
     * The scope of this secret. The secret will only be usable for jobs in this
     * particular scope.
     */
    scope: ExecutorSecretScope

    /**
     * If true, this secret is defined in a namespace and a secret with the same key
     * is also defined in the global namespace, which this secret overwrites.
     */
    overwritesGlobalSecret: boolean

    /**
     * The namespace this secret belongs to. Null, if a global secret.
     * Global secrets are available to every execution.
     */
    namespace: Namespace | null

    /**
     * The creator of the secret. Null, if the creator has been deleted.
     */
    creator: IUser | null

    /**
     * The date and time this secret has been created.
     */
    createdAt: DateTime

    /**
     * The date and time this secret has been last updated.
     */
    updatedAt: DateTime

    /**
     * The list of access events to this secret. Every time the secret value is
     * decoded and used, one of these entries is created.
     */
    accessLogs: IExecutorSecretAccessLogConnection
}

export interface IAccessLogsOnExecutorSecretArguments {
    /**
     * Only return N records.
     * @default 50
     */
    first?: number | null

    /**
     * Opaque cursor for pagination.
     */
    after?: string | null
}

/**
 * A list of executor secrets.
 */
export interface IExecutorSecretConnection {
    __typename: 'ExecutorSecretConnection'

    /**
     * A list of executor secrets.
     */
    nodes: IExecutorSecret[]

    /**
     * The total number of records in this result set.
     */
    totalCount: number

    /**
     * Pagination information.
     */
    pageInfo: IPageInfo
}
// ...
export interface IPageInfo {
    __typename: 'PageInfo'

    /**
     * When paginating forwards, the cursor to continue.
     */
    endCursor: string | null

    /**
     * When paginating forwards, are there more items?
     */
    hasNextPage: boolean
}
// ...
export interface IUserConnection {
    __typename: 'UserConnection'

    /**
     * A list of users.
     */
    nodes: IUser[]

    /**
     * The total count of users in the connection. This total count may be larger
     * than the number of nodes in this object when the result is paginated.
     */
    totalCount: number

    /**
     * Pagination information.
     */
    pageInfo: IPageInfo
}
// ...
export interface IOrg {
    __typename: 'Org'

    /**
     * The unique ID for the organization.
     */
    id: ID

    /**
     * The organization's name. This is unique among all organizations on this Sourcegraph site.
     */
    name: string

    /**
     * The organization's chosen display name.
     */
    displayName: string | null

    /**
     * The date when the organization was created.
     */
    createdAt: DateTime

    /**
     * A list of users who are members of this organization.
     */
    members: IUserConnection

    /**
     * The latest settings for the organization.
     * Only organization members and site admins can access this field.
     */
    latestSettings: ISettings | null

    /**
     * All settings for this organization, and the individual levels in the settings cascade (global > organization)
     * that were merged to produce the final merged settings.
     * Only organization members and site admins can access this field.
     */
    settingsCascade: ISettingsCascade

    /**
     * DEPRECATED
     * @deprecated "Use settingsCascade instead. This field is a deprecated alias for it and will be removed in a future release."
     */
    configurationCascade: IConfigurationCascade

    /**
     * DEPRECATED
     * A pending invitation for the viewer to join this organization, if any.
     * @deprecated "Use invitationByToken operation instead. This field is deprecated and will be removed in a future release."
     */
    viewerPendingInvitation: IOrganizationInvitation | null

    /**
     * Whether the viewer has admin privileges on this organization. Currently, all of an organization's members
     * have admin privileges on the organization.
     */
    viewerCanAdminister: boolean

    /**
     * Whether the viewer is a member of this organization.
     */
    viewerIsMember: boolean

    /**
     * The URL to the organization.
     */
    url: string

    /**
     * The URL to the organization's settings.
     */
    settingsURL: string | null

    /**
     * The name of this user namespace's component. For organizations, this is the organization's name.
     */
    namespaceName: string

    /**
     * A list of batch changes initially applied in this organization.
     */
    batchChanges: IBatchChangeConnection

    /**
     * The list of all available executor secrets for execution in this orgs namespace.
     */
    executorSecrets: IExecutorSecretConnection
}
// ...
export interface IOrganizationInvitation {
    __typename: 'OrganizationInvitation'

    /**
     * The ID of the invitation.
     */
    id: ID

    /**
     * The organization that the invitation is for.
     */
    organization: IOrg

    /**
     * The user who sent the invitation.
     */
    sender: IUser

    /**
     * The user who received the invitation.
     */
    recipient: IUser | null

    /**
     * The email address that the invitation was sent to.
     */
    recipientEmail: string | null

    /**
     * The date when this invitation was created.
     */
    createdAt: DateTime

    /**
     * The most recent date when a notification was sent to the recipient about this invitation.
     */
    notifiedAt: DateTime | null

    /**
     * The date when this invitation was responded to by the recipient.
     */
    respondedAt: DateTime | null

    /**
     * The recipient's response to this invitation, or no response (null).
     */
    responseType: OrganizationInvitationResponseType | null

    /**
     * The URL where the recipient can respond to the invitation when pending, or null if not pending.
     */
    respondURL: string | null

    /**
     * The date when this invitation was revoked.
     */
    revokedAt: DateTime | null

    /**
     * The date when this invitation is going to expire.
     */
    expiresAt: DateTime | null

    /**
     * Boolean flag which returns true if the email on the invite matches a verified email of the user
     */
    isVerifiedEmail: boolean | null
}
// ...
export enum OrganizationInvitationResponseType {
    /**
     * The invitation was accepted by the recipient.
     */
    ACCEPT = 'ACCEPT',

    /**
     * The invitation was rejected by the recipient.
     */
    REJECT = 'REJECT',
}
// ...
export interface IDefaultSettings {
    __typename: 'DefaultSettings'

    /**
     * The opaque GraphQL ID.
     */
    id: ID

    /**
     * The latest default settings (this never changes).
     */
    latestSettings: ISettings | null

    /**
     * The URL to the default settings. This URL does not exist because you
     * cannot edit or directly view default settings.
     */
    settingsURL: string | null

    /**
     * Whether the viewer can modify the subject's settings. Always false for
     * default settings.
     */
    viewerCanAdminister: boolean

    /**
     * The default settings, and the final merged settings.
     * All viewers can access this field.
     */
    settingsCascade: ISettingsCascade

    /**
     * DEPRECATED
     * @deprecated "Use settingsCascade instead. This field is a deprecated alias for it and will be removed in a future release."
     */
    configurationCascade: IConfigurationCascade
}
// ...
export type SettingsSubject = IUser | IOrg | IDefaultSettings | ISite

/**
 * SettingsSubject is something that can have settings: a site ("global settings", which is different from "site
 * configuration"), an organization, or a user.
 */
export interface ISettingsSubject {
    __typename: 'SettingsSubject'

    /**
     * The ID.
     */
    id: ID

    /**
     * The latest settings.
     */
    latestSettings: ISettings | null

    /**
     * The URL to the settings.
     */
    settingsURL: string | null

    /**
     * Whether the viewer can modify the subject's settings.
     */
    viewerCanAdminister: boolean

    /**
     * All settings for this subject, and the individual levels in the settings cascade (global > organization > user)
     * that were merged to produce the final merged settings.
     */
    settingsCascade: ISettingsCascade

    /**
     * DEPRECATED
     * @deprecated "Use settingsCascade instead. This field is a deprecated alias for it and will be removed in a future release."
     */
    configurationCascade: IConfigurationCascade
}

/**
 * The configurations for all of the relevant settings subjects, plus the merged settings.
 */
export interface ISettingsCascade {
    __typename: 'SettingsCascade'

    /**
     * The other settings subjects that are applied with lower precedence than this subject to
     * form the final merged settings. For example, a user in 2 organizations would have the following
     * settings subjects: site (global settings), org 1, org 2, and the user.
     */
    subjects: SettingsSubject[]

    /**
     * The effective final merged settings as (stringified) JSON, merged from all of the subjects.
     */
    final: string

    /**
     * DEPRECATED: This field will be removed in a future release.
     * The effective final merged settings, merged from all of the subjects.
     * @deprecated "use final instead"
     */
    merged: IConfiguration
}

/**
 * DEPRECATED: Renamed to SettingsCascade.
 */
export interface IConfigurationCascade {
    __typename: 'ConfigurationCascade'

    /**
     * DEPRECATED
     * @deprecated "use SettingsCascade.subjects instead"
     */
    subjects: SettingsSubject[]

    /**
     * DEPRECATED
     * @deprecated "use SettingsCascade.final instead"
     */
    merged: IConfiguration
}

/**
 * Settings is a version of a configuration settings file.
 */
export interface ISettings {
    __typename: 'Settings'

    /**
     * The ID.
     */
    id: number

    /**
     * The subject that these settings are for.
     */
    subject: SettingsSubject

    /**
     * The author, or null if there is no author or the authoring user was deleted.
     */
    author: IUser | null

    /**
     * The time when this was created.
     */
    createdAt: DateTime

    /**
     * The stringified JSON contents of the settings. The contents may include "//"-style comments and trailing
     * commas in the JSON.
     */
    contents: JSONCString

    /**
     * DEPRECATED: This field will be removed in a future release.
     * The configuration.
     * @deprecated "use the contents field instead"
     */
    configuration: IConfiguration
}

/**
 * DEPRECATED: Use the contents field on the parent type instead. This type will be removed in a future release.
 */
export interface IConfiguration {
    __typename: 'Configuration'

    /**
     * DEPRECATED: This field will be removed in a future release.
     * The raw JSON contents, encoded as a string.
     * @deprecated "use the contents field on the parent type instead"
     */
    contents: JSONCString

    /**
     * DEPRECATED: This field is always empty. It will be removed in a future release.
     * @deprecated "use client-side JSON Schema validation instead"
     */
    messages: string[]
}